* Change View procedure complete.  ==> The implemented change view doesn't check change view digests and doesn't handle missed requests during new-view construction procedure.  ==> Change view timer doesn't check request arrival order.

+++ jds/src/br/ufba/lasid/jds/prototyping/hddss/LoadAwareCPU.java
@@ -20,7 +20,7 @@ public class LoadAwareCPU extends CPU{
       
       try{
          int dsize = XObject.objectToByteArray(data).length * 8;
-         delay = dsize / getProcessingRate();
+         delay = ((double)dsize / (double)getProcessingRate());
       }catch(Exception e){
          
       }
